done error handlings

+++ db/db_user.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm.session import Session
 
 from db.hash import Hash
@@ -22,11 +23,22 @@ def get_all_users(db: Session):
 
 
 def get_user(db: Session, id: int):
-    return db.query(DBUser).filter(DBUser.id == id).first()
+    user = db.query(DBUser).filter(DBUser.id == id).first()
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"The user with id {id} not found",
+        )
+    return user
 
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DBUser).filter(DBUser.id == id)
+    if not user.first():
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"The user with id {id} not found",
+        )
     user.update(
         {
             DBUser.username: request.username,
@@ -40,6 +52,11 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DBUser).filter(DBUser.id == id).first()
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"The user with id {id} not found",
+        )
     db.delete(user)
     db.commit()
     return "user deleted"
